% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegusta.R
\docType{package}
\name{eegusta-package}
\alias{eegusta}
\alias{eegusta-package}
\title{EEG Utilities and STatistical Analysis: eeguana-based tools for statistical analysis of EEG data}
\description{
EEG Utilities and STatistical Analysis: eeguana-based tools for statistical analysis of eeg data. It contains a few utiliy functions and some examples of import, preprocessing and Statistical Analysis of EEG data (see Vignettes folder). This project is a work in progress.
}
\author{
Livio Finos, Antonio Maffei, Angela Andreella
}
\keyword{eeg}
