% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegUtils2eeguana.R
\name{eegUtils2eeguana}
\alias{eegUtils2eeguana}
\title{Converts eeg_data or eeg_epochs (eegUtils) to eeg_lst (eeguana) object}
\usage{
eegUtils2eeguana(data, reference, unit, resolution)
}
\arguments{
\item{data}{eegUtils data}

\item{reference}{reference}

\item{unit}{unit}

\item{resolution}{resolution}
}
\value{
Returns an eeg_lst (eeguana package)
}
\description{
Builds an eeg_lst object composed of two data.table objects and a tibble one. All three are linked by a unique identifier .id. Amplitude values and timestamps appear in the signal table. Triggers, blinks, artifact rejection markings, and other events logged by the EEG recording software appear in the events table. Segment information and recording IDs appear in the segments tibble.
}
\author{
Angela Andreella
}
