% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim2eeg_data.R
\name{read_delim2eeg_data}
\alias{read_delim2eeg_data}
\title{Reads aplitudes from a text delimited file}
\usage{
read_delim2eeg_data(
  file,
  time_column = "time",
  srate = 256,
  chan_info = NULL,
  .unit = "milliseconds"
)
}
\arguments{
\item{file}{file's name}

\item{time_column}{name of the column that contains the time}

\item{srate}{sampling rate. 256 is the default}

\item{chan_info}{a \code{tibble} with channel locations. It has the following columns \code{electrode, radius, theta, phi, cart_x, cart_y, cart_z, x, y}(\code{NULL} by default)}

\item{.unit}{if "milliseconds" it divides the time by 1000}
}
\value{
returns a tibble object with columns \code{.channel,radius,theta,phi,.x,.y,.z}.
}
\description{
Reads aplitudes from a text delimited file
}
\author{
Livio Finos
}
