% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_elp.R
\name{read_elp}
\alias{read_elp}
\title{Reads channel locations in elp format}
\usage{
read_elp(file)
}
\arguments{
\item{file}{file's name}
}
\value{
returns a tibble object with columns \code{.channel,radius,theta,phi,.x,.y,.z}.
}
\description{
Reads channel locations in elp format
}
