% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegusta.R
\docType{package}
\name{eegusta-package}
\alias{eegusta}
\alias{eegusta-package}
\title{eegusta: EEGUana+STAtistics: eeguana-based tools for statistical analysis of eegdata}
\description{
EEGUana+STAtistics: eeguana-based tools for statistical analysis of eegdata.
}
\keyword{internal}
